feat: initial commit for refactor
